split into layers
