+++ data.json
@@ -1,72 +1,1 @@
-{
-  "77c349a6-fa12-40f1-837d-289ea88c9786": {
-    "name": "Test",
-    "id": "77c349a6-fa12-40f1-837d-289ea88c9786",
-    "column": "To Do"
-  },
-  "5831d139-6af3-4c60-92c6-f8e606f1ee8d": {
-    "name": "Mega Test",
-    "id": "5831d139-6af3-4c60-92c6-f8e606f1ee8d",
-    "column": "To Do"
-  },
-  "98d25147-dab1-445e-bb58-ab6eab33d04a": {
-    "name": "asdf",
-    "id": "98d25147-dab1-445e-bb58-ab6eab33d04a",
-    "column": "To Do"
-  },
-  "0d1d982a-07b0-4c97-8ba1-7cf6e1a7a1ae": {
-    "name": "Hello world",
-    "id": "0d1d982a-07b0-4c97-8ba1-7cf6e1a7a1ae",
-    "column": "To Do"
-  },
-  "542ef6ee-36d0-44aa-bd9b-4936e608af2c": {
-    "name": "TEasting",
-    "id": "542ef6ee-36d0-44aa-bd9b-4936e608af2c",
-    "column": "To Do"
-  },
-  "56d9f076-5d5d-4723-bfbc-04ce279a1971": {
-    "name": "Tasks",
-    "id": "56d9f076-5d5d-4723-bfbc-04ce279a1971",
-    "column": "To Do"
-  },
-  "38a22b2b-6867-46e9-b6fd-a3fa298056b0": {
-    "name": "Refactor",
-    "id": "38a22b2b-6867-46e9-b6fd-a3fa298056b0",
-    "column": "To Do"
-  },
-  "90764725-a6ac-420d-aba9-81da391f9762": {
-    "name": "Testytest",
-    "id": "90764725-a6ac-420d-aba9-81da391f9762",
-    "column": "To Do"
-  },
-  "3d5b4d82-fdf5-487b-809a-63462df78bcb": {
-    "name": "In progress...",
-    "id": "3d5b4d82-fdf5-487b-809a-63462df78bcb",
-    "column": "Doing"
-  },
-  "e72d8fbe-7e71-4fc9-9162-98526bb8b015": {
-    "name": "Finished this one",
-    "id": "e72d8fbe-7e71-4fc9-9162-98526bb8b015",
-    "column": "Done"
-  },
-  "bf6366a1-4476-4727-a481-ba831017c707": {
-    "name": "And this one",
-    "id": "bf6366a1-4476-4727-a481-ba831017c707",
-    "column": "Done"
-  },
-  "e2a506b3-7ed2-418d-b17f-915416751e70": {
-    "name": "Crikey,",
-    "id": "e2a506b3-7ed2-418d-b17f-915416751e70",
-    "column": "Done"
-  },
-  "0fa54cc3-3a10-4c84-9ed5-17da14c0f471": {
-    "name": "Wasn't I...",
-    "id": "0fa54cc3-3a10-4c84-9ed5-17da14c0f471",
-    "column": "Done"
-  },
-  "d46c713e-f9ad-474f-b6ca-06a08caf9f2a": {
-    "name": "Productive!",
-    "id": "d46c713e-f9ad-474f-b6ca-06a08caf9f2a",
-    "column": "Done"
-  }
-}
+{"77c349a6-fa12-40f1-837d-289ea88c9786":{"name":"Test","id":"77c349a6-fa12-40f1-837d-289ea88c9786","column":"To Do"},"5831d139-6af3-4c60-92c6-f8e606f1ee8d":{"name":"Mega Test","id":"5831d139-6af3-4c60-92c6-f8e606f1ee8d","column":"To Do"},"98d25147-dab1-445e-bb58-ab6eab33d04a":{"name":"asdf","id":"98d25147-dab1-445e-bb58-ab6eab33d04a","column":"To Do"},"0d1d982a-07b0-4c97-8ba1-7cf6e1a7a1ae":{"name":"Hello world","id":"0d1d982a-07b0-4c97-8ba1-7cf6e1a7a1ae","column":"To Do"},"542ef6ee-36d0-44aa-bd9b-4936e608af2c":{"name":"TEasting","id":"542ef6ee-36d0-44aa-bd9b-4936e608af2c","column":"To Do"},"56d9f076-5d5d-4723-bfbc-04ce279a1971":{"name":"Tasks","id":"56d9f076-5d5d-4723-bfbc-04ce279a1971","column":"To Do"},"38a22b2b-6867-46e9-b6fd-a3fa298056b0":{"name":"Refactor","id":"38a22b2b-6867-46e9-b6fd-a3fa298056b0","column":"To Do"},"90764725-a6ac-420d-aba9-81da391f9762":{"name":"Testytest","id":"90764725-a6ac-420d-aba9-81da391f9762","column":"Done"},"3d5b4d82-fdf5-487b-809a-63462df78bcb":{"name":"In progress...","id":"3d5b4d82-fdf5-487b-809a-63462df78bcb","column":"Doing"},"e72d8fbe-7e71-4fc9-9162-98526bb8b015":{"name":"Finished this one","id":"e72d8fbe-7e71-4fc9-9162-98526bb8b015","column":"Done"},"bf6366a1-4476-4727-a481-ba831017c707":{"name":"And this one","id":"bf6366a1-4476-4727-a481-ba831017c707","column":"Done"},"e2a506b3-7ed2-418d-b17f-915416751e70":{"name":"Crikey,","id":"e2a506b3-7ed2-418d-b17f-915416751e70","column":"Done"},"0fa54cc3-3a10-4c84-9ed5-17da14c0f471":{"name":"Wasn't I...","id":"0fa54cc3-3a10-4c84-9ed5-17da14c0f471","column":"Done"},"d46c713e-f9ad-474f-b6ca-06a08caf9f2a":{"name":"Productive!","id":"d46c713e-f9ad-474f-b6ca-06a08caf9f2a","column":"Done"}}
