+++ main.ts
@@ -1,5 +1,7 @@
 // @ts-types="npm:@types/express"
-import { randomUUID, UUID } from "node:crypto";
+import {  UUID } from "node:crypto";
+
+import { getBoard, addTask, updateTask, removeTask } from "./service.ts";
 
 import express from "npm:express";
 
@@ -14,72 +16,6 @@ app.use(express.json());
 app.get("/", function (_req, res) {
   res.render("pages/index");
 });
-
-type Task = {
-  name: string;
-  id: UUID;
-  column: string;
-};
-type Column = {
-  name: string;
-  tasks: Array<Task>;
-};
-
-type TaskMap = Record<UUID, Task>;
-interface ColumnMap {
-  [key: string]: Column;
-}
-async function writeTasks(tasks: TaskMap) {
-  await Deno.writeTextFile("./data.json", JSON.stringify(tasks));
-}
-
-async function readTasks() {
-  const data = await Deno.readTextFile("./data.json");
-  
-  return JSON.parse(data);
-}
-
-
-async function getBoard(){
-  const output:ColumnMap = {}
-  const taskMap:TaskMap = await readTasks();
-
-  Object.values(taskMap).forEach((task) => {
-    const { column } = task;
-    if (!output[column]) {
-      output[column] = { name: column, tasks: [] };
-    }
-    output[column].tasks.push(task);
-  });
-
-  return Object.values(output);
-  
-}
-
-async function updateTask(id: UUID, name: string, column: string) {
-  
-  const taskMap = await readTasks();
-
-  taskMap[id].name = name
-  taskMap[id].column = column
-
-  await writeTasks(taskMap);
-}
-
-async function removeTask(id: UUID) {
-  const taskMap = await readTasks();
-
-  taskMap.delete(id)
-
-  await writeTasks(taskMap);
-}
-
-async function addTask( name: string, column: string){
-  const newTask:Task = { name, id: randomUUID(), column };
-  const taskMap = await readTasks();
-  taskMap[newTask.id]=newTask
-  await writeTasks(taskMap);
-}
 
 app.post("/tasks", async (req, res) => {
   const taskName = req.body.taskName;
@@ -119,7 +55,6 @@ if (import.meta.main) {
 
 app.get("/board", async function (_req, res) {
   const columns = await getBoard();
-  console.log(columns)
   res.render("pages/board", { columns });
 });
